// src/app/interfaces/paciente.ts
export interface PacienteInterface {
    id: string;
    folio: string;
    curp: string;
    fnacimiento: string;
    enacimiento: string;
    norigen: string;
    cp: string;
    ereside:string;
    mun: string;
    col: string;
    sexo: string;
    nombre:string
    apaterno:string
    amaterno:string
    telefono:string
    correo:string
}